Extract selector generation logic
