+++ lorm-macros/src/orm/by.rs
@@ -1,7 +1,5 @@
 use crate::models::{OrmModel, PrimaryKey};
-use crate::utils::{
-    db_placeholder, get_bind_param_type_and_usage, get_bind_type_where_constraint,
-};
+use crate::utils::{db_placeholder, get_bind_param_type_and_usage, get_bind_type_where_constraint};
 use quote::{__private::TokenStream, format_ident, quote};
 use syn::spanned::Spanned;
 
@@ -16,7 +14,7 @@ pub fn generate_by(
     let table_name = &model.table_name;
     let all_columns = model.full_select_columns();
 
-    let stream: Vec<(TokenStream, TokenStream)> = model.fields.iter().filter(|field| field.column_properties.generate_by).map(|field| {
+    let stream: Vec<(TokenStream, TokenStream)> = model.fields.iter().filter(|field| field.should_generate_selector(&model.primary_key)).map(|field| {
         let lifetime = quote! {'a};
         let parameter = quote! {value};
         let field_type_constraints = get_bind_type_where_constraint(&field.ty, database_type, &lifetime).unwrap();
@@ -52,7 +50,7 @@ pub fn generate_by(
             PrimaryKey::Generated(..) => None.into_iter(),
             PrimaryKey::Manual(fields) => {
                 if fields.len() > 1 {
-                    let by_fn_name = &model.primary_key_by_name;
+                    let by_fn_name = &model.primary_key_selector;
                     let lifetime = quote! {'a};
 
                     let type_constraints = fields
